third_party/blink/web_tests/platform/generic/fast/forms/color/input-color-choose-default-value-after-set-value-expected.txt

Test if change event fires when the user selects the default value after the value was changed by JS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.value is "#ff0000"
PASS input.value is "#ffffff"
onchange fired: #ffffff
PASS successfullyParsed is true

TEST COMPLETE
